requirements.txt
aiohttp==3.7.2
pyTelegramBotAPI==3.7.3
python-telegram-bot==13.0

